requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
psycopg2-binary
pydantic
requests
python-multipart
alembic
apscheduler
firebase-admin
pyOpenSSL